+++ src/main/java/jumper/filter/RequestFilter.java
@@ -215,8 +215,7 @@ public class RequestFilter extends AbstractGatewayFilterFactory<RequestFilter.Co
                             TokenInfo tokenInfo =
                                 oauthTokenUtil.getAccessTokenWithOauthCredentialsObject(
                                     jumperConfig.getExternalTokenEndpoint(),
-                                    oauthCredentials.get(),
-                                    jumperConfig.getConsumer());
+                                    oauthCredentials.get());
 
                             HeaderUtil.addHeader(
                                 exchange,
@@ -293,6 +292,7 @@ public class RequestFilter extends AbstractGatewayFilterFactory<RequestFilter.Co
                       log.info("logging request: {}", value("jumperInfo", infoRequest));
                     });
 
+                HeaderUtil.removeHeaders(exchange, jumperConfig.getRemoveHeaders());
                 tracer.currentSpan().event("jrqf");
               });
 
@@ -378,7 +378,7 @@ public class RequestFilter extends AbstractGatewayFilterFactory<RequestFilter.Co
     if (Objects.nonNull(clientId) && Objects.nonNull(clientSecret)) {
       TokenInfo tokenInfo =
           oauthTokenUtil.getAccessTokenWithClientCredentials(
-              tokenEndpoint, clientId, clientSecret, clientScope, consumer);
+              tokenEndpoint, clientId, clientSecret, clientScope);
       HeaderUtil.addHeader(
           exchange,
           Constants.HEADER_AUTHORIZATION,