+++ src/main/java/jumper/model/config/OauthCredentials.java
@@ -12,11 +12,13 @@ import lombok.Setter;
 public class OauthCredentials {
   private String clientId;
   private String clientSecret;
+  private String clientKey;
   private String scopes;
   private String username;
   private String password;
   private String refreshToken;
   private String grantType;
+  private String tokenRequest;
 
   public String getId() {