+++ src/main/java/jumper/service/OauthTokenUtil.java
@@ -4,38 +4,27 @@
 
 package jumper.service;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import static jumper.Constants.TOKEN_REQUEST_METHOD_POST;
+
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.Header;
 import io.jsonwebtoken.Jwt;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.MalformedJwtException;
-import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.UnsupportedJwtException;
 import io.jsonwebtoken.security.SignatureException;
 import io.netty.channel.ConnectTimeoutException;
 import io.netty.handler.ssl.SslHandshakeTimeoutException;
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
+import java.util.*;
 import java.util.concurrent.*;
 import jumper.Constants;
 import jumper.model.TokenInfo;
 import jumper.model.config.JumperConfig;
-import jumper.model.config.KeyInfo;
 import jumper.model.config.OauthCredentials;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.security.oauth2.core.AuthorizationGrantType;
@@ -59,20 +48,8 @@ public class OauthTokenUtil {
 
   private final WebClient oauthTokenUtilWebClient;
   private final TokenCacheService tokenCache;
+  private final TokenGeneratorService tokenGenerator;
   private final BasicAuthUtil basicAuthUtil;
-
-  private static String securityPath;
-  private static String securityFile;
-
-  @Value("${jumper.security.dir:keypair}")
-  private void setSecurityPath(String name) {
-    securityPath = name;
-  }
-
-  @Value("${jumper.security.file:private.json}")
-  private void setSecurityFile(String name) {
-    securityFile = name;
-  }
 
   public static String getTokenWithoutSignature(String consumerToken) {
 
@@ -178,7 +155,7 @@ public class OauthTokenUtil {
       claims.put(Constants.TOKEN_CLAIM_AUD, aud);
     }
 
-    return generateToken(claims, issuer, expiration, issuedAt, jc.getRealmName());
+    return tokenGenerator.fromRealm(claims, issuer, expiration, issuedAt, jc.getRealmName());
   }
 
   public String generateGatewayTokenForPublisher(String issuer, String realm) {
@@ -187,7 +164,7 @@ public class OauthTokenUtil {
     claims.put(Constants.TOKEN_CLAIM_AZP, "stargate");
     claims.put(Constants.TOKEN_CLAIM_CLIENT_ID, "gateway");
 
-    return generateToken(
+    return tokenGenerator.fromRealm(
         claims,
         issuer,
         new Date(System.currentTimeMillis() + 300 * 1000),
@@ -195,68 +172,18 @@ public class OauthTokenUtil {
         realm);
   }
 
-  private String generateToken(
-      HashMap<String, String> claims, String issuer, Date expiration, Date issuedAt, String realm) {
-    Map<String, KeyInfo> keyInfoMap;
-
-    try {
-      log.debug("GatewayToken or OneToken: Loading keyInfo");
-      keyInfoMap = loadKeyInfo();
-
-    } catch (IOException e1) {
-      log.error("IOException", e1);
-      throw new RuntimeException("Error while generating LMS token, key info missing");
-    }
-
-    if (!keyInfoMap.containsKey(realm)) {
-      throw new RuntimeException("key info missing for realm " + realm);
-    }
-
-    return Jwts.builder()
-        .setClaims(claims)
-        .setIssuer(issuer)
-        .setExpiration(expiration)
-        .setIssuedAt(issuedAt)
-        .signWith(keyInfoMap.get(realm).getPk(), SignatureAlgorithm.RS256)
-        .setHeaderParam("kid", keyInfoMap.get(realm).getKid())
-        .setHeaderParam("typ", "JWT")
-        .compact();
-  }
-
-  public static Map<String, KeyInfo> loadKeyInfo() throws IOException {
-    Path kidFile =
-        Path.of(
-            System.getProperty("user.dir")
-                + File.separator
-                + securityPath
-                + File.separator
-                + securityFile);
-
-    TypeReference<HashMap<String, KeyInfo>> typeRef = new TypeReference<>() {};
-
-    return new ObjectMapper()
-        .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
-        .readValue(Files.readString(kidFile), typeRef);
-  }
-
   public TokenInfo getInternalMeshAccessToken(JumperConfig jc) {
     return getAccessTokenWithClientCredentials(
         jc.getInternalTokenEndpoint() + Constants.ISSUER_SUFFIX,
         jc.getClientId(),
         jc.getClientSecret(),
-        null,
-        "");
+        null);
   }
 
   public TokenInfo getAccessTokenWithClientCredentials(
-      String tokenEndpoint,
-      String clientID,
-      String clientSecret,
-      String scope,
-      String subscriberClientId) {
+      String tokenEndpoint, String clientID, String clientSecret, String scope) {
 
-    final String tokenKey =
-        tokenCache.generateTokenCacheKey(tokenEndpoint, clientID, subscriberClientId);
+    final String tokenKey = tokenCache.generateTokenCacheKey(tokenEndpoint, clientID, scope);
 
     // try to get valid token from tokenCache...
     return tokenCache
@@ -279,11 +206,9 @@ public class OauthTokenUtil {
   }
 
   public TokenInfo getAccessTokenWithOauthCredentialsObject(
-      String tokenEndpoint, OauthCredentials oauthCredentials, String subscriberClientId) {
+      String tokenEndpoint, OauthCredentials oauthCredentials) {
 
-    final String tokenKey =
-        tokenCache.generateTokenCacheKey(
-            tokenEndpoint, oauthCredentials.getId(), subscriberClientId);
+    final String tokenKey = tokenCache.generateTokenCacheKey(tokenEndpoint, oauthCredentials);
 
     // try to get valid token from tokenCache...
     return tokenCache
@@ -293,12 +218,33 @@ public class OauthTokenUtil {
               MultiValueMap<String, String> requestParameter = new LinkedMultiValueMap<>();
               String basicAuth = null;
 
+              if (StringUtils.isNotBlank(oauthCredentials.getClientKey())) {
+                requestParameter.add(
+                    Constants.TOKEN_REQUEST_PARAMETER_CLIENT_ID, oauthCredentials.getClientId());
+                requestParameter.add(
+                    Constants.TOKEN_REQUEST_PARAMETER_CLIENT_ASSERTION,
+                    createJwtTokenForExternalIdp(tokenEndpoint, oauthCredentials));
+                requestParameter.add(
+                    Constants.TOKEN_REQUEST_PARAMETER_CLIENT_ASSERTION_TYPE,
+                    Constants.TOKEN_REQUEST_PARAMETER_CLIENT_ASSERTION_TYPE_JWT);
+              }
+
               if (StringUtils.isNotBlank(oauthCredentials.getClientId())
                   && StringUtils.isNotBlank(oauthCredentials.getClientSecret())) {
 
-                basicAuth =
-                    basicAuthUtil.encodeBasicAuth(
-                        oauthCredentials.getClientId(), oauthCredentials.getClientSecret());
+                if (StringUtils.isNotBlank(oauthCredentials.getTokenRequest())
+                    && StringUtils.equalsIgnoreCase(
+                        TOKEN_REQUEST_METHOD_POST, oauthCredentials.getTokenRequest())) {
+                  requestParameter.add(
+                      Constants.TOKEN_REQUEST_PARAMETER_CLIENT_ID, oauthCredentials.getClientId());
+                  requestParameter.add(
+                      Constants.TOKEN_REQUEST_PARAMETER_CLIENT_SECRET,
+                      oauthCredentials.getClientSecret());
+                } else {
+                  basicAuth =
+                      basicAuthUtil.encodeBasicAuth(
+                          oauthCredentials.getClientId(), oauthCredentials.getClientSecret());
+                }
               }
 
               if (StringUtils.isNotBlank(oauthCredentials.getUsername())
@@ -326,6 +272,29 @@ public class OauthTokenUtil {
 
               return getAccessTokenQuery(tokenEndpoint, tokenKey, requestParameter, basicAuth);
             });
+  }
+
+  private String createJwtTokenForExternalIdp(
+      String tokenEndpoint, OauthCredentials oauthCredentials) {
+    /*
+    iss - REQUIRED. Issuer. This MUST contain the client_id of the OAuth Client.
+    sub - REQUIRED. Subject. This MUST contain the client_id of the OAuth Client.
+    aud - REQUIRED. Audience. The aud (audience) Claim. Value that identifies the Authorization Server as an intended audience. The Authorization Server MUST verify that it is an intended audience for the token. The Audience SHOULD be the URL of the Authorization Server's Token Endpoint.
+    jti - REQUIRED. JWT ID. A unique identifier for the token, which can be used to prevent reuse of the token. These tokens MUST only be used once, unless conditions for reuse were negotiated between the parties; any such negotiation is beyond the scope of this specification.
+    exp - REQUIRED. Expiration time on or after which the JWT MUST NOT be accepted for processing.
+    iat - OPTIONAL. Time at which the JWT was issued.
+    */
+    HashMap<String, String> claims = new HashMap<>();
+    claims.put(Constants.TOKEN_CLAIM_SUB, oauthCredentials.getClientId());
+    claims.put(Constants.TOKEN_CLAIM_AUD, tokenEndpoint);
+    claims.put(Constants.TOKEN_CLAIM_JTI, UUID.randomUUID().toString());
+
+    return tokenGenerator.fromKey(
+        claims,
+        oauthCredentials.getClientId(),
+        new Date(System.currentTimeMillis() + 60 * 1000),
+        new Date(System.currentTimeMillis()),
+        oauthCredentials.getClientKey());
   }
 
   private TokenInfo getAccessTokenQuery(