+++ src/test/java/jumper/mocks/MockIrisServer.java
@@ -18,13 +18,16 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 import jumper.model.TokenInfo;
 import jumper.util.AccessToken;
+import lombok.extern.slf4j.Slf4j;
 import org.mockserver.client.MockServerClient;
 import org.mockserver.integration.ClientAndServer;
 import org.mockserver.model.Header;
 import org.mockserver.model.HttpError;
+import org.mockserver.model.RegexBody;
 import org.springframework.http.HttpHeaders;
 import org.springframework.util.Base64Utils;
 
+@Slf4j
 public class MockIrisServer {
 
   private ClientAndServer mockServer;
@@ -33,7 +36,7 @@ public class MockIrisServer {
 
   private final String irisLocalHost = "localhost";
 
-  private static int responseCode = 200;
+  private int responseCode = 200;
 
   public void startServer() {
     mockServer = startClientAndServer(irisLocalPort);
@@ -46,7 +49,7 @@ public class MockIrisServer {
   public void createExpectationInternalToken(String id) {
 
     String tokenInfoJson = getTokenInfoJson(CONSUMER_GATEWAY);
-    List<Header> headersList = getHeaderList("86");
+    List<Header> headersList = getHeaderList();
 
     new MockServerClient(irisLocalHost, irisLocalPort)
         .when(
@@ -80,6 +83,37 @@ public class MockIrisServer {
                 .withBody(
                     addIdSuffix("client_id=external_configured", id)
                         + "&client_secret=secret&grant_type=client_credentials"),
+            exactly(1))
+        .respond(
+            response()
+                .withStatusCode(responseCode)
+                .withHeaders(
+                    new Header("Content-Type", "application/json; charset=utf-8"),
+                    new Header("Cache-Control", "no-store"))
+                .withBody(tokenInfoJson)
+                .withDelay(TimeUnit.SECONDS, 1));
+  }
+
+  public void createExpectationExternalTokenKeyed(String id) {
+
+    String tokenInfoJson = getTokenInfoJson(CONSUMER_EXTERNAL_CONFIGURED);
+
+    new MockServerClient(irisLocalHost, irisLocalPort)
+        .when(
+            request()
+                .withMethod("POST")
+                .withPath("/external")
+                .withBody(
+                    RegexBody.regex(
+                        ".*("
+                            + addIdSuffix("client_id=external_configured", id)
+                            + "|"
+                            + "client_assertion=eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9."
+                            + "|"
+                            + "client_assertion_type=urn:ietf:params:oauth:client-assertion-type:jwt-bearer"
+                            + "|"
+                            + "grant_type=client_credentials"
+                            + ")+")),
             exactly(1))
         .respond(
             response()
@@ -252,10 +286,11 @@ public class MockIrisServer {
                     new Header("Content-Type", "application/json; charset=utf-8"),
                     new Header("Cache-Control", "no-store"))
                 .withBody(
-                    "{\n"
-                        + "\t\"error\": \"unauthorized_client\",\n"
-                        + "\t\"error_description\": \"Invalid client or Invalid client credentials\"\n"
-                        + "}"));
+                    """
+								{
+								\t"error": "unauthorized_client",
+								\t"error_description": "Invalid client or Invalid client credentials"
+								}"""));
   }
 
   public void createExpectationDropConnection(String id) {
@@ -361,11 +396,11 @@ public class MockIrisServer {
         .respond(response().withStatusCode(responseCode).withDelay(TimeUnit.SECONDS, 1));
   }
 
-  private List<Header> getHeaderList(String contentLength) {
+  private List<Header> getHeaderList() {
     List<Header> headersList = new ArrayList<>();
     headersList.add(new Header(HttpHeaders.HOST, irisLocalHost + ":" + irisLocalPort));
     headersList.add(new Header(HttpHeaders.ACCEPT, "*/*"));
-    headersList.add(new Header(HttpHeaders.CONTENT_LENGTH, contentLength));
+    headersList.add(new Header(HttpHeaders.CONTENT_LENGTH, "86"));
     headersList.add(
         new Header(HttpHeaders.CONTENT_TYPE, "application/x-www-form-urlencoded;charset=UTF-8"));
     return headersList;
@@ -387,7 +422,7 @@ public class MockIrisServer {
     try {
       tokenInfoJson = mapper.writeValueAsString(tokenInfo);
     } catch (JsonProcessingException e) {
-      e.printStackTrace();
+      log.error(e.getMessage());
     }
     return tokenInfoJson;
   }