+++ src/test/java/jumper/util/AccessToken.java
@@ -17,7 +17,6 @@ import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 import jumper.Constants;
 import jumper.model.config.KeyInfo;
-import jumper.service.OauthTokenUtil;
 import lombok.Builder;
 
 @Builder
@@ -61,7 +60,7 @@ public class AccessToken {
     Date expiration = new Date(System.currentTimeMillis() + TimeUnit.MINUTES.toMillis(5));
     KeyInfo keyInfo = null;
     try {
-      keyInfo = OauthTokenUtil.loadKeyInfo().get(Constants.DEFAULT_REALM);
+      keyInfo = jumper.service.TokenGeneratorService.loadKeyInfo().get(Constants.DEFAULT_REALM);
     } catch (IOException e) {
       e.getStackTrace();
     }